Move ReportPanic to cmdutil and report panic metric
